Feature/invalid note reason

+++ app/models/entry.rb
@@ -1,7 +1,9 @@
 class Entry < ActiveRecord::Base
   include Reportable
 
-  attr_accessible :image, :live_demo_url, :idea_url, :name, :description, :member_id, :url, :technologies, :letter_under_oath, :repo_url, :demo_url
+  attr_accessible :image, :live_demo_url, :idea_url, :name, :description,
+                  :member_id, :url, :technologies, :letter_under_oath,
+                  :repo_url, :demo_url, :invalid_reason
 
   belongs_to :member
   belongs_to :challenge
@@ -59,11 +61,19 @@ class Entry < ActiveRecord::Base
 
   def mark_as_valid!
     self.is_valid = true
+    self.invalid_reason = nil
+    self.save!
   end
 
-  def mark_as_invalid!
-    self.report_cards.destroy_all
+  def mark_as_invalid!(message)
+    # we need a reason for this invalid_reason to be blank
+    return false if message.blank?
+    # destroy all report cards that belong to this entry
+    # 'cause it has now been marked as invalid
     self.is_valid = false
+    self.invalid_reason = message
+    self.save!
+    self.report_cards.destroy_all
   end
 
   def publish!
